add spotify module
